Move UUID tests to tests/types/test_uuid.py
The UUID cases were the last remaining slice of the
`test_default_validators` parameterization, which is now removed
entirely together with the `CheckModel` fixture.
